20180416-1.5.1(21): 1. 修复参数配置导入因无法正确解析测量量类型而失败的问题

+++ app/src/main/java/com/weisi/tool/wsnbox/io/database/SensorDatabase.java
@@ -906,7 +906,7 @@ public class SensorDatabase implements Constant {
                     mIntoMeasurementElement = true;
                     String index = attributes.getValue("index");
                     mMeasurementValueId = LogicalSensor.ID.getId(mAddress,
-                            Byte.parseByte(attributes.getValue(TAG_TYPE), 16),
+                            (byte) Integer.parseInt(attributes.getValue(TAG_TYPE), 16),
                             TextUtils.isEmpty(index) ? 0 : Integer.parseInt(index));
                     break;
                 case TAG_WARNER:
